działające filtrowanie i obrazki :)

+++ routes/web.php
@@ -17,19 +17,21 @@ use App\Http\Controllers\WelcomeController;
 */
 Route::get('/',[WelcomeController::class, 'index']);
 
-Route::middleware(['auth', 'verified'])->group(function() {
-    Route::middleware(['can:isAdmin'])->group(function() {
-        Route::resource('obozy', OboztController::class);
-
-        Route::get('/users/list', [UserController::class, 'index']);
-        Route::delete('/users/{user}', [UserController::class, 'destroy']);
-    });
+Route::get('/obozy',[ObozController::class, 'index'])->name('obozy.index')->middleware('auth') ;
+Route::get('/obozy/create',[ObozController::class, 'create'])->name('obozy.create')->middleware('auth') ;
+Route::get('/obozy/{oboz}',[ObozController::class, 'show'])->name('obozy.show')->middleware('auth') ;
+Route::post('/obozy',[ObozController::class, 'store'])->name('obozy.store')->middleware('auth') ;
+Route::get('/obozy/edit/{oboz}',[ObozController::class, 'edit'])->name('obozy.edit')->middleware('auth') ;
+Route::post('/obozy/{oboz}',[ObozController::class, 'update'])->name('obozy.update')->middleware('auth') ;
+Route::delete('/obozy/{oboz}',[ObozController::class, 'destroy'])->name('obozy.destroy')->middleware('auth');   
+    
     Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-});
 
 
+
+Route::get('/users/list',[UserController::class, 'index'])->middleware('auth') ;
+Route::delete('/users/{user}',[UserController::class, 'destroy'])->middleware('auth');
 
 Auth::routes();
-
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/',[WelcomeController::class, 'index']);
